+++ src/NosCore.Data/WebApi/ConnectedAccount.cs
@@ -25,7 +25,7 @@ namespace NosCore.Data.WebApi
     {
         public string? Name { get; set; }
         public RegionType Language { get; set; }
-        public int ChannelId { get; set; }
+        public long ChannelId { get; set; }
         public Character? ConnectedCharacter { get; set; }
     }
 }
